Sketch html+tar combination prototyped earlier
There idea here is that contrary to the pure WASM module we get to
bootstrap a page with normal web tools; and then utilize the web
assembly thing as a bonus that is conveniently attached to such a page
and can conveniently consume the extra tar contents. The synergy of
having an initial file system as required for a WASI environment is the
main speculated benefit of the combination.

With this we may de-prioritize other targets and in particular those
without any HTML structure. There are more synergy effects in the
combination. For instance, we may run the web assembly module once for
some set elements, e.g. run a module implementing the `dot` rendering
process or mermaid-js equivalents in parallel mapped over specifying
elements. This would, in particular, generalize to allowing the DOM to
specify an initial process graph in interesting and hopefully convenient
ways. The interpretation of elements to select program binaries and
environments arguments is up of the environment of course (with a vague
idea of having a micro-kernel addition to WASI in some instances).