Customization point for stage2: wasi-loader
Different environments may encode their data payloads in different ways.
The customization point introduced here is rather suboptimal in the
sense that it is not transparent. The difference will be visible to the
stage 3 code which gets the data as zip in one case and a prepared
filesystem in the other.

Indeed creating the zip at runtime calls for (rather) heavy javascript
dependencies that would be dead-weight in other environments. Although
the size may be defensible. The unzip stage 3 wasm-boot containing the
inverse may be more problematic and will not be optional if we maintain
a strict sense of layering (where inputs are normalized at some point).

A more modular approach may be to change the primary unzipping data
approach. Instead of the WASI module itself a more accessible special
file may be created which it is filled with the WASI module in on case
and a synthesized archive in another.